enquiry update API

+++ Home-Loan-Management-System/src/main/java/in/cjctech/wonderhomeapp/app/repository/EnquiryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import in.cjctech.wonderhomeapp.app.model.Enquiry;
 
 @Repository
-public interface EnquiryRepository extends JpaRepository<Enquiry, Integer> {
+public interface EnquiryRepository extends JpaRepository<Enquiry, Long> {
 
 }
